+++ eras/conway/impl/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename `ConwayPParams` to be consistent with the Agda specification. #
   * `govActionExpiration` to `govActionLifetime`
   * `committeeTermLimit` to `committeeMaxTermLength`
+  * `minCommitteeSize` to `committeeMinSize`
 * Prevent `DRep` expiry when there are no active governance proposals to vote on (in ConwayCERTS). #3729
   * Add `updateNumDormantEpochs` function in `ConwayEPOCH` to update the dormant-epochs counter
   * Refactor access to `ConwayGovState` by making its lens part of `ConwayEraGov`.